Factor out AccessMode to security package
Move CatalogAccessControlRule::AccessMode to
security package
